Move Faraday_Proxy_Tools/FaradayIO to faraday/proxyio
This patch creates a top-level faraday module which begins with the
proxy I/O library. Scripts are updated to import via the new module
path. Scripts which are expected to be executed from the command line
have a Python #! line added and are marked executable.
