+++ Applications/Hermes/hermesflaskserver.py
@@ -1,5 +1,6 @@
+#!/usr/bin/env python
+
 import os
-import hermesobject as hermesobject
 import ConfigParser
 import json
 import base64
@@ -7,6 +8,11 @@ import sys
 
 from flask import Flask
 from flask import request
+
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+
+import hermesobject
 
 # Open configuration INI
 config = ConfigParser.RawConfigParser()
