+++ Applications/Hermes/hermesmessage.py
@@ -1,14 +1,8 @@
 # imports
 import struct
-import sys
-import os
 
-sys.path.append(os.path.join(os.path.dirname(__file__),
-                             "../../Faraday_Proxy_Tools"))  # Append path to common tutorial FaradayIO module
-# noinspection PyPep8
-from FaradayIO import faradaybasicproxyio
-# noinspection PyPep8
-from FaradayIO import faradaycommands
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
 
 
 class MsgStateMachineTx(object):
