+++ Applications/deviceconfiguration/readconfig.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 # Warning - Must run the "deviceconfiguration" proxy application
 
 # readconfig.py is a simple script that only reads/displays the current programmed flash configuration without
@@ -14,13 +16,13 @@ import time
 import ConfigParser
 import json
 
-sys.path.append(os.path.join(os.path.dirname(__file__),
-                             "../../Faraday_Proxy_Tools"))  # Append path to common tutorial FaradayIO module
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
 
 # Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import telemetryparser
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import telemetryparser
 
 # Open configuration INI
 config = ConfigParser.RawConfigParser()
