+++ Tutorials/Python_Developer_Tutorials/foundation/Commanding-Local/Proxy-Local-Commands-Tutorial.py
@@ -1,20 +1,21 @@
+#!/usr/bin/env python
+
 # This example will show how to command Faraday using the basic command application.
 # LED's being commanded may be receiving other commands and not work as intended (i.e. RED due to RF TX indication)
 
 # Imports - General
-
 import os
 import sys
 import time
 import ConfigParser
 
-sys.path.append(os.path.join(os.path.dirname(__file__),
-                             "../../../../Faraday_Proxy_Tools"))  # Append path to common tutorial FaradayIO module
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../../..'))
 
 # Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import gpioallocations
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import gpioallocations
 
 # Open configuration INI
 config = ConfigParser.RawConfigParser()
