+++ Tutorials/Python_Developer_Tutorials/foundation/Commanding-Remote-RF/Tutorial_Remote_Command.py
@@ -1,15 +1,18 @@
-#Imports - General
+#!/usr/bin/env python
 
+#Imports - General
 import os
 import sys
+import time
 import ConfigParser
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../../../Faraday_Proxy_Tools"))  #Append path to common tutorial FaradayIO module
+
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../../..'))
 
 #Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import gpioallocations
-import time
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import gpioallocations
 
 
 #Setup a Faraday IO object
