+++ Tutorials/Python_Developer_Tutorials/foundation/Telemetry-Parsing/Tutorial_Telemetry_Parsing.py
@@ -1,14 +1,18 @@
+#!/usr/bin/env python
+
 #Imports - General
 
 import os
 import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../../../Faraday_Proxy_Tools"))  #Append path to common tutorial FaradayIO module
 
-#Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import telemetryparser
-from FaradayIO import cc430radioconfig
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../../..'))
+
+# Imports - Faraday Specific
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import telemetryparser
+from faraday.proxyio import cc430radioconfig
 
 
 #Variables
