+++ Tutorials/Python_Developer_Tutorials/miscellaneous/Device-Configuration/Tutorial_Device_Configuration.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 #Warning - Must run the "deviceconfiguration" proxy application
 
 #Imports - General
@@ -9,12 +10,13 @@ import base64
 import cPickle
 import time
 
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../../../Faraday_Proxy_Tools"))  #Append path to common tutorial FaradayIO module
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../../..'))
 
-#Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import telemetryparser
+# Imports - Faraday Specific
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import telemetryparser
 
 #Variables
 local_device_callsign = 'REPLACEME'  # Should match the connected Faraday unit as assigned in Proxy configuration
