+++ Tutorials/Python_Developer_Tutorials/miscellaneous/Factory-Reset-Configuration/Factory_Reset.py
@@ -1,15 +1,17 @@
-#Imports - General
+#!/usr/bin/env python
 
+#Imports - General
 import os
 import sys
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../../../Faraday_Proxy_Tools"))  #Append path to common tutorial FaradayIO module
-
-#Imports - Faraday Specific
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
-from FaradayIO import telemetryparser
-
 import time
+
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../../../..'))
+
+# Imports - Faraday Specific
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
+from faraday.proxyio import telemetryparser
 
 
 #Variables
