+++ Tutorials/start/hello-world.py
@@ -1,10 +1,15 @@
+#!/usr/bin/env python
+
 import os
 import sys
 import time
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../Faraday_Proxy_Tools"))  #Append path to common tutorial FaradayIO module
 
-from FaradayIO import faradaybasicproxyio
-from FaradayIO import faradaycommands
+# Add Faraday library to the Python path.
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+
+# Imports - Faraday Specific
+from faraday.proxyio import faradaybasicproxyio
+from faraday.proxyio import faradaycommands
 
 #Setup a Faraday IO object
 faraday_1 = faradaybasicproxyio.proxyio()  #default proxy port
