+++ docs/conf.py
@@ -18,8 +18,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../Faraday_Proxy_Tools/FaradayIO'))  #FaradayIO toolset
-sys.path.insert(0, os.path.abspath('../Python_Developer_Tutorials/FaradayIO/Tutorial_0'))  #FaradayIO toolset
+sys.path.insert(0, os.path.abspath('../faraday'))  # Faraday library
 
 # -- General configuration ------------------------------------------------
